requirements.txt
python-telegram-bot==22.5
pdfplumber==0.11.8
google-generativeai==0.8.5
pillow==10.4.0
pandas==2.3.3
pytesseract==0.3.13
opencv-python==4.8.1.78
python-dotenv==1.0.1
numpy==1.26.4
requests==2.31.0
cryptography==43.0.3
httpx==0.27.0
sniffio==1.3.1
ta==0.11.0
yfinance==0.2.66
scikit-learn==1.7.2
reportlab==4.0.4
pandas-datareader==0.10.0
scipy==1.13.1
